measurements ready

+++ Automated_testing/cuda_test/execute.py
@@ -20,91 +20,71 @@ from SLtools import *
 
 cases = import_cases(path = "./cases.csv",type = "cuda")
 
-
-
 case_number = len(cases)
 case_count = get_case_number(path = "./case_number.cache")
-
-#set_case_number(5)
-
 
 print("Cases: " , case_number)
 print("Start case id: " , case_count)
 
-
-
-
-quit()
-
 #temp_file_name = "temp_script_file.txt"
 
-
+#creating names for files
 measurement_name = "800_CUDA"
 measurement_csv = measurement_name + "_results.csv"
 measurement_summary = measurement_name + "_summary.log"
 
 #so,thread_limit,iteration,x,y,z,execution time, begin temperature,end temperature
 
-os.popen("module list >> {}".format(measurement_summary)).read()
-os.popen("date >> {}".format(measurement_summary)).read()
+# do this in the slurm job script
+#os.popen("module list >> {}".format(measurement_summary)).read()
+#os.popen("date >> {}".format(measurement_summary)).read()
 
+f = None
+if os.path.exists(measurement_csv):
+	pass
+	#f = open(measurement_csv, "a")
+else:
+	f = open(measurement_csv, "w")
+	f.write("so,thread_limit,size,x,y,z,execution time,begin temperature,end temperature\n")
+	f.close()
 
-f = open(measurement_csv, "w")
-f.write("so,thread_limit,size,x,y,z,execution time,begin temperature,end temperature\n")
-f.close()
-
+#init folders
 os.popen("make clean").read()
 os.popen("make init").read()
 
 start_time = time.time()
 
-for so in space_orders:
+#if needed do controll cases here
+
+
+for i in range(case_count,len(cases)):
+	case = cases[i]
+	area,so,x,y,z,thread_limit = case
+	print("Case: %d/%d" % (case_count,case_number))
+	begin_temp = get_gpu_temperature()
+	print("temp:",begin_temp,"C")
+	extra = "-DF{}_{} -Dblocksize_x={} -Dblocksize_y={} -Dblocksize_z={} -DTHREADLIMIT={}".format(area,so,x,y,z,thread_limit)
+	#print(extra)
+	build_main(extra)
+	forward_time = 	test_run()
+	end_temp = get_gpu_temperature()
+	#print("forward took:",forward_time,"temperature",end_temp)
 	
-	#begin_temp = get_gpu_temperature()
-	for idx in sizes:
-		
-		x,y,z,thread_limit = idx
-		print("Case: %d/%d" % (case_count,case_number))
+	f = open(measurement_csv, "a")
+	f.write("%d,%d,%d,%d,%d,%d,%s,%d,%d\n" % (so,thread_limit,area,x,y,z,forward_time,begin_temp,end_temp))
+	f.close()
+	
+	ellapsed = time.time()-start_time
+	
+	#print(ellapsed/60,"m ellapsed,",(ellapsed/case_count)*(case_number-case_count)/60,"m left")
+	
+	wait_for_gpu_60celsius()
+	print("----------------------------------")
+	case_count+=1
+	set_case_number(case_count)
+	#quit()
+	#break
 
-
-		case_count+=1
-		begin_temp = get_gpu_temperature()
-		print("temperature",begin_temp)
-
-		extra = "-DF{}_{} -Dblocksize_x={} -Dblocksize_y={} -Dblocksize_z={} -DTHREADLIMIT".format(area,so,x,y,z,t)
-
-		#print(extra)
-		build_main(extra)
-		forward_time = 	test_run()
-
-		end_temp = get_gpu_temperature()
-
-		print("forward took:",forward_time,"temperature",end_temp)
-		f = open(measurement_csv, "a")
-		f.write("%d,%d,%d,%d,%d,%d,%s,%d,%d\n" % (so,thread_limit,area,x,y,z,forward_time,begin_temp,end_temp))
-		f.close()
-
-		ellapsed = time.time()-start_time
-		print(ellapsed/60,"m ellapsed,",(ellapsed/case_count)*(case_number-case_count)/60,"m left")
-
-
-		while True:
-			temp_t = get_gpu_temperature()
-			if temp_t>60:
-				#for sec in range(5):
-				#print("Temperature:",temp_t," C",end="\r")
-				time.sleep(5)
-				#print("60 sec cooling complete.")
-			else:
-				print("60 C  reached.")
-				break
-		print("----------------------------------")
-		#quit()
-		#break
-
-
-
-os.popen("date >> {}".format(measurement_summary)).read()
 print("----------finished?------------")
